Migrate ContainsVertexWithTests to swift-testing
Related to https://github.com/ikelax/swift-dependency-graphs/issues/15
